thumbloaded gallery pics now do the proper javascript opening in iframe not opening in new page.
I made the customscript that controlled the javascript click registering into a function called window.clickifyGallery() and then called it from the thumbloaderscript when a new image is added. I also crashed my computer by forgetting not to use a ';' in coffeescript.... Yeah.

also I deleted comentform which I accidentally made a while back. You may notice that comment is not spelt coment.