Several grammar corrections made to about.html
Almost all of these are capitalization errors, it's important to take care of them to look as professional as possible. I also removed the words "or get bought" in "if rhombik ever goes down or gets bought" because a big statement of yours is that you're not going to give in and sell (a platform that isn't going to be bought by your competitors.

I also think you should throw in some talk to Makers as well because it seems like your main statement is directed towards bigger companies, which are not who you should be focusing at the moment.