Cleaned up Comments and Info section
Links posted in the Comments or the Readme.md section of download pages now are bolded and a color (#66357F) between the top bar's color and blue, and purple (#800800) when already visited.

Also fixed the border to Readme.md, the tab had one but the actual information box was missing one. Need to find a way to get rid of the border between readme.md and the box.