Fixed the same title bug! It seems I have already been passing data on a project to the form (useful) so what happens is now:
In the clean function in the project form:
If the project.title == cleaned_data["title"] ## the new and old title are the same
    cleaned_data["title"] = "" ## the new title is set to blank

Then when the form data is being saved from the project view:
If form.cleaned_data["title"]: ## if there is a new title
    project.title = form.cleaned_data["tilte"] ## the new title is saved

Otherwise the old title is used. This does have the side effect that the user can set the title to blank and the old title will be used, but I don't see this as a bug so much as a feature.