projects from thingi importer check validity on spider closing now! The user just has to click publish.
There is a new project method called enf_validity which is called
if you call .save(enf_valid=True)
There is a new project method called get_form() which returns a form
filled in with the data for the project. It's used in the edit form and
in the thingi importer.